gck: Fast 2-D Walsh-Hadamard transform using GCK.
Currently a bit of a mess. Includes 3 ways of doing the DC
computation: with integral images, a two-pass convolution,
and a single-pass convolution.

The convolved signal is expanded to N + M - 1 elements for
the x, y directions. Integral image retains the original image
dimensions.

Also includes a bunch of print/test/debug code for now.