gck: Consider width of target when interleaving bases.
The target width is not always the same as the number of bases.
An example is the luma and chroma planes of an image. Here
you can only interleave one color plane at a time, so the
actual width of the target is larger when taking other planes into
consideration.