kdtree: Calculate x,y coords for target in source.
This avoids relying on the xy node field which has proven to be
problematic. For example, in case of identical tuples, if they
are in different nodes (say, one is a an inner node and another
is a leaf), then only the inner node will get assigned a xy value
during the get_positions phase. If the leaf gets selected during
matching, it has an invalid xy coord, and BAM! segfault.

So instead, since the coeffs are stored in the original order
with the new quickselect (only pointers into the coeffs are
permuted), we can calculate the index of the tuple by finding its
offset from the beginning of the array, and the x/y from there.

This guarantees that every tuple will have an x/y coordinate.

Also un-macroify some stuff that was done in the process of
debugging this, which I'm too lazy to put back together. Probably
best that way, anyway.